Fixed correlation function and regression function

+++ include/statsi/statsi.hpp
@@ -1,7 +1,6 @@
 #ifndef STATSI_HPP
 #define STATSI_HPP
 
-
 #include <cmath>
 #include <vector>
 
@@ -9,12 +8,11 @@ using namespace std;
 float calculateMedian(vector<float> values);
 double calculateSum(vector<float>& values);
 float calculateMean(vector<float>& values);
-
 float calculateVariance(vector<float>& values);
 float calculateStandardDeviation(vector<float>& values);
 float calculateCorrelation(vector<float>& col1_values,
                            vector<float>& col2_values);
-tuple<float, float> calculateRegression(vector<float>& values);
+tuple<double, double> calculateRegression(vector<float>& values);
 float getMax(vector<float>& values);
 float getMin(vector<float>& values);
 bool isPrime(int number);
